+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/server.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/bufbuild/connect-go"
 	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/resources"
+	"github.com/bufbuild/connect-go"
 
 	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/helm"
 
@@ -32,6 +32,7 @@ import (
 	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/paginate"
 	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/pkgutils"
 	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/resourcerefs"
+	"github.com/SAP/kubeapps/cmd/kubeapps-apis/plugins/pkg/safelog"
 	log "k8s.io/klog/v2"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -166,12 +167,12 @@ func NewServer(configGetter core.KubernetesConfigGetter, kubeappsCluster string,
 // state. For the fluxv2 plugin:
 //   - if flux helm-controller flag "-no-cross-namespace-refs=true" is
 //     enabled only the request target namespace is relevant
-//     ref https://github.com/vmware-tanzu/kubeapps/issues/5541
+//     ref https://github.com/SAP/kubeapps/issues/5541
 //   - otherwise the request context namespace (the target
 //     namespace) is not relevant since charts from a repository in any namespace
 //     accessible to the user are available to be installed in the target namespace.
 func (s *Server) GetAvailablePackageSummaries(ctx context.Context, request *connect.Request[corev1.GetAvailablePackageSummariesRequest]) (*connect.Response[corev1.GetAvailablePackageSummariesResponse], error) {
-	log.Infof("+fluxv2 GetAvailablePackageSummaries(request: [%v])", request)
+	safelog.Request("+fluxv2 GetAvailablePackageSummaries", request)
 	defer log.Info("-fluxv2 GetAvailablePackageSummaries")
 
 	// grpc compiles in getters for you which automatically return a default (empty) struct
@@ -206,7 +207,7 @@ func (s *Server) GetAvailablePackageSummaries(ctx context.Context, request *conn
 		return nil, err
 	}
 
-	// per https://github.com/vmware-tanzu/kubeapps/pull/3686#issue-1038093832
+	// per https://github.com/SAP/kubeapps/pull/3686#issue-1038093832
 	for _, summary := range packageSummaries {
 		summary.AvailablePackageRef.Context.Cluster = s.kubeappsCluster
 	}
@@ -230,7 +231,7 @@ func (s *Server) GetAvailablePackageSummaries(ctx context.Context, request *conn
 
 // GetAvailablePackageDetail returns the package metadata managed by the 'fluxv2' plugin
 func (s *Server) GetAvailablePackageDetail(ctx context.Context, request *connect.Request[corev1.GetAvailablePackageDetailRequest]) (*connect.Response[corev1.GetAvailablePackageDetailResponse], error) {
-	log.Infof("+fluxv2 GetAvailablePackageDetail(request: [%v])", request)
+	safelog.Request("+fluxv2 GetAvailablePackageDetail", request)
 	defer log.Info("-fluxv2 GetAvailablePackageDetail")
 
 	if request == nil || request.Msg.AvailablePackageRef == nil {
@@ -260,7 +261,7 @@ func (s *Server) GetAvailablePackageDetail(ctx context.Context, request *connect
 
 // GetAvailablePackageVersions returns the package versions managed by the 'fluxv2' plugin
 func (s *Server) GetAvailablePackageVersions(ctx context.Context, request *connect.Request[corev1.GetAvailablePackageVersionsRequest]) (*connect.Response[corev1.GetAvailablePackageVersionsResponse], error) {
-	log.Infof("+fluxv2 GetAvailablePackageVersions [%v]", request)
+	safelog.Request("+fluxv2 GetAvailablePackageVersions", request)
 	defer log.Info("-fluxv2 GetAvailablePackageVersions")
 
 	if request.Msg.GetPkgVersion() != "" {
@@ -302,7 +303,7 @@ func (s *Server) GetAvailablePackageVersions(ctx context.Context, request *conne
 
 // GetInstalledPackageSummaries returns the installed packages managed by the 'fluxv2' plugin
 func (s *Server) GetInstalledPackageSummaries(ctx context.Context, request *connect.Request[corev1.GetInstalledPackageSummariesRequest]) (*connect.Response[corev1.GetInstalledPackageSummariesResponse], error) {
-	log.Infof("+fluxv2 GetInstalledPackageSummaries [%v]", request)
+	safelog.Request("+fluxv2 GetInstalledPackageSummaries", request)
 	itemOffset, err := paginate.ItemOffsetFromPageToken(request.Msg.GetPaginationOptions().GetPageToken())
 	if err != nil {
 		return nil, err
@@ -336,7 +337,7 @@ func (s *Server) GetInstalledPackageSummaries(ctx context.Context, request *conn
 
 // GetInstalledPackageDetail returns the package metadata managed by the 'fluxv2' plugin
 func (s *Server) GetInstalledPackageDetail(ctx context.Context, request *connect.Request[corev1.GetInstalledPackageDetailRequest]) (*connect.Response[corev1.GetInstalledPackageDetailResponse], error) {
-	log.Infof("+fluxv2 GetInstalledPackageDetail [%v]", request)
+	safelog.Request("+fluxv2 GetInstalledPackageDetail", request)
 
 	if request == nil || request.Msg.InstalledPackageRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request InstalledPackageRef provided"))
@@ -366,7 +367,7 @@ func (s *Server) GetInstalledPackageDetail(ctx context.Context, request *connect
 
 // CreateInstalledPackage creates an installed package based on the request.
 func (s *Server) CreateInstalledPackage(ctx context.Context, request *connect.Request[corev1.CreateInstalledPackageRequest]) (*connect.Response[corev1.CreateInstalledPackageResponse], error) {
-	log.Infof("+fluxv2 CreateInstalledPackage [%v]", request)
+	safelog.Request("+fluxv2 CreateInstalledPackage", request)
 
 	if request == nil || request.Msg.AvailablePackageRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request AvailablePackageRef provided"))
@@ -410,7 +411,7 @@ func (s *Server) CreateInstalledPackage(ctx context.Context, request *connect.Re
 
 // UpdateInstalledPackage updates an installed package based on the request.
 func (s *Server) UpdateInstalledPackage(ctx context.Context, request *connect.Request[corev1.UpdateInstalledPackageRequest]) (*connect.Response[corev1.UpdateInstalledPackageResponse], error) {
-	log.Infof("+fluxv2 UpdateInstalledPackage [%v]", request)
+	safelog.Request("+fluxv2 UpdateInstalledPackage", request)
 
 	if request == nil || request.Msg.InstalledPackageRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request InstalledPackageRef provided"))
@@ -439,7 +440,7 @@ func (s *Server) UpdateInstalledPackage(ctx context.Context, request *connect.Re
 
 // DeleteInstalledPackage deletes an installed package.
 func (s *Server) DeleteInstalledPackage(ctx context.Context, request *connect.Request[corev1.DeleteInstalledPackageRequest]) (*connect.Response[corev1.DeleteInstalledPackageResponse], error) {
-	log.Infof("+fluxv2 DeleteInstalledPackage [%v]", request)
+	safelog.Request("+fluxv2 DeleteInstalledPackage", request)
 
 	if request == nil || request.Msg.InstalledPackageRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request InstalledPackageRef provided"))
@@ -519,7 +520,7 @@ func (s *Server) AddPackageRepository(ctx context.Context, request *connect.Requ
 }
 
 func (s *Server) GetPackageRepositoryDetail(ctx context.Context, request *connect.Request[corev1.GetPackageRepositoryDetailRequest]) (*connect.Response[corev1.GetPackageRepositoryDetailResponse], error) {
-	log.Infof("+fluxv2 GetPackageRepositoryDetail [%v]", request)
+	safelog.Request("+fluxv2 GetPackageRepositoryDetail", request)
 	defer log.Info("-fluxv2 GetPackageRepositoryDetail")
 	if request == nil || request.Msg.PackageRepoRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request AvailablePackageRef provided"))
@@ -548,7 +549,7 @@ func (s *Server) GetPackageRepositoryDetail(ctx context.Context, request *connec
 
 // GetPackageRepositorySummaries returns the package repositories managed by the 'fluxv2' plugin
 func (s *Server) GetPackageRepositorySummaries(ctx context.Context, request *connect.Request[corev1.GetPackageRepositorySummariesRequest]) (*connect.Response[corev1.GetPackageRepositorySummariesResponse], error) {
-	log.Infof("+fluxv2 GetPackageRepositorySummaries [%v]", request)
+	safelog.Request("+fluxv2 GetPackageRepositorySummaries", request)
 	cluster := request.Msg.GetContext().GetCluster()
 	if cluster != "" && cluster != s.kubeappsCluster {
 		return nil, connect.NewError(connect.CodeUnimplemented, fmt.Errorf("not supported yet: request.Context.Cluster: [%v]", cluster))
@@ -565,7 +566,7 @@ func (s *Server) GetPackageRepositorySummaries(ctx context.Context, request *con
 
 // UpdatePackageRepository updates a package repository based on the request.
 func (s *Server) UpdatePackageRepository(ctx context.Context, request *connect.Request[corev1.UpdatePackageRepositoryRequest]) (*connect.Response[corev1.UpdatePackageRepositoryResponse], error) {
-	log.Infof("+fluxv2 UpdatePackageRepository [%v]", request)
+	safelog.Request("+fluxv2 UpdatePackageRepository", request)
 	if request == nil || request.Msg.PackageRepoRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request PackageRepoRef provided"))
 	}
@@ -587,7 +588,7 @@ func (s *Server) UpdatePackageRepository(ctx context.Context, request *connect.R
 
 // DeletePackageRepository deletes a package repository based on the request.
 func (s *Server) DeletePackageRepository(ctx context.Context, request *connect.Request[corev1.DeletePackageRepositoryRequest]) (*connect.Response[corev1.DeletePackageRepositoryResponse], error) {
-	log.Infof("+fluxv2 DeletePackageRepository [%v]", request)
+	safelog.Request("+fluxv2 DeletePackageRepository", request)
 	if request == nil || request.Msg.PackageRepoRef == nil {
 		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("no request PackageRepoRef provided"))
 	}
@@ -606,7 +607,7 @@ func (s *Server) DeletePackageRepository(ctx context.Context, request *connect.R
 }
 
 func (s *Server) GetPackageRepositoryPermissions(ctx context.Context, request *connect.Request[corev1.GetPackageRepositoryPermissionsRequest]) (*connect.Response[corev1.GetPackageRepositoryPermissionsResponse], error) {
-	log.Infof("+fluxv2 GetPackageRepositoryPermissions [%v]", request)
+	safelog.Request("+fluxv2 GetPackageRepositoryPermissions", request)
 
 	cluster := request.Msg.GetContext().GetCluster()
 	namespace := request.Msg.GetContext().GetNamespace()