+++ cmd/kubeapps-apis/plugins/helm/packages/v1alpha1/getresourcerefs_test.go
@@ -67,7 +67,7 @@ func TestGetInstalledPackageResourceRefs(t *testing.T) {
 		newTestCase(3, "my-apache", true, 0),
 		newTestCase(4, "my-iis", false, connect.CodeNotFound),
 		newTestCase(5, "my-apache", false, connect.CodeInternal),
-		// See https://github.com/vmware-tanzu/kubeapps/issues/632
+		// See https://github.com/SAP/kubeapps/issues/632
 		newTestCase(6, "my-apache", true, 0),
 		newTestCase(7, "my-apache", true, 0),
 		newTestCase(8, "my-apache", true, 0),